refactor: split offline test suites into feature-organized submodules
Split the two monolithic offline test files into feature-organized
submodules for easier navigation and maintenance.

CLI offline (64 tests): 13 submodules (usage, issues, teams, projects,
documents, cycles, milestones, embeds, comments, auth, self_update,
format, labels).

SDK offline (39 tests): 2 submodules (queries, mutations).

Uses #[path] attributes since Rust integration tests don't follow
standard edition 2021 module resolution for subdirectories.
