+++ crates/lineark/tests/offline/labels.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[test]
+fn labels_list_help_shows_team_flag() {
+    lineark()
+        .args(["labels", "list", "--help"])
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("--team"));
+}
